QuizXMento - A Quiz App/requirements.txt
Flask==2.0.1
Flask-MySQLdb==1.0.1
pandas==1.3.3
openpyxl==3.0.9
